fix(cli): enforce CLI separation by delegating to backing modules (fixes #210)
Refactors code.py, plan.py, and init.py to be thin CLI wrappers that
delegate all business logic to engine.run, graph.planner, and
workspace.init_project respectively. Inverts the dependency direction
so backing modules never import from CLI modules. Updates all test
mock paths accordingly.
