+++ tests/unit/session/test_steering.py
@@ -103,7 +103,7 @@ class TestLoadSteeringReturnNoneForPlaceholderOnly:
 
     def test_returns_none_for_placeholder(self, tmp_path: Path) -> None:
         """load_steering() returns None when file contains only the placeholder."""
-        from agent_fox.cli.init import _ensure_steering_md
+        from agent_fox.workspace.init_project import _ensure_steering_md
         from agent_fox.session.prompt import load_steering
 
         _ensure_steering_md(tmp_path)
